FINALLY, A BALANCED STICK ;D
So much time has been put into this stick balancing reinforcement learner. It uses a Q-learning model to balance the stick. It's not 100 percent perfect because it doesn't know the cart speed and only has 2 speed states in each direction for the angular rotation of the stick. Some improvements could definitely be made, but I'm definitely satisfied with this model. Now to use this base code to make the first version of the creatures that learn to walk!