FIRST SIMULATED CREATURE!!!
I just knocked out all of the required algorithims and challenges that I needed to add to make a simulated creature that learns to walk using Q-learning. I don't believe Q-learning is appropriate for this application though, because the state space is really pushing it to the limit here. I mean this has over 5000 different states with 9 different actions for each state, and that's in the simplest form of a walking creature! I definitely need to learn how to implement deep-q-learning if I want to take this any further. I'm pretty satisfied with the results, I set up the environment to reward forward velocity and altitude, so the A.I. took advantage of the auto-flip functionality to get into the air. It was pretty cool watching it, but after I implement the deep learning functionality, I'm going straight to building the UI and the fully fleshed out Terrain-Adaptive Kinetic creatures!