ReLu Neural Network functionality finished
I should, heavy emphasis on SHOULD, be able to dynamically create and train a neural network of any size now. This was a huge pain to build because there was just so much logic and math involved. I had never even built a ReLu neural network before this! Also, I added a ton of convenience extension functions that really clean up the code with the matrix operations.